refractor: extract generate manifest
Refractor server API codes for unit testing #2
