+++ lib/manifestPlugin.js
@@ -1,5 +1,5 @@
 const sharp = require('sharp')
-const rename = require('rename')
+const generateManifest = require('./generateManifest')
 
 const ID = 'gridsome:manifest-plugin'
 
@@ -13,53 +13,27 @@ module.exports = class ManifestPlugin {
       const logger = compiler.getInfrastructureLogger(ID)
 
       compilation.hooks.additionalAssets.tapPromise(ID, async () => {
-        const options = this.options
-        const manifestDest = options.manifestPath
-        const iconsDir = 'assets/icons/'
-        const iconName = options.icon.split('/').slice(-1)[0]
-        const type = 'image/' + iconName.split('.').slice(-1)[0]
-
-        // Generate all size images from options.icon
-        const sizes = [512, 384, 192, 180, 152, 144, 128, 120, 96, 72, 48, 16]
-
-        const icons = []
+        const manifestDest = this.options.manifestPath
+        const outputManifest = generateManifest(this.options)
+        const manifestString = JSON.stringify(outputManifest)
+        logger.info('Writing manifest to ' + manifestDest)
+        compilation.emitAsset(manifestDest, {
+          source: () => manifestString,
+          size: () => manifestString.length
+        })
         logger.info('Generating icons')
         await Promise.all(
-          sizes.map(async (size) => {
-            const sizes = `${size}x${size}`
-            const src = iconsDir + rename(iconName, { suffix: `-${sizes}` })
-            icons.push({
-              src,
-              type,
-              sizes,
-              purpose: options.maskableIcon ? 'maskable any' : 'any'
-            })
-            const imageData = await sharp(options.icon)
+          outputManifest.icons.map(async icon => {
+            const size = parseInt(icon.sizes.match(/\d+/), 10)
+            const imageData = await sharp(this.options.icon)
               .resize(size, size)
               .toBuffer()
-            compilation.emitAsset(src, {
+            compilation.emitAsset(icon.src, {
               source: () => imageData,
               size: () => imageData.length
             })
           })
         )
-        const defaultManifest = {
-          start_url: '.',
-          display: 'standalone',
-          background_color: '#000000',
-          icons,
-          name: options.name,
-          short_name: options.name,
-          theme_color: options.themeColor
-        }
-        const outputManifest = JSON.stringify(
-          Object.assign(defaultManifest, options.manifestOptions)
-        )
-        logger.info('Writing manifest to ' + manifestDest)
-        compilation.emitAsset(manifestDest, {
-          source: () => outputManifest,
-          size: () => outputManifest.length
-        })
       })
     })
   }
